seeds finish
